Move ptyHostMain into the node folder
