+++ src/vs/platform/terminal/node/ptyService.ts
