+++ src/vs/workbench/buildfile.desktop.js
@@ -28,7 +28,7 @@ exports.collectModules = function () {
 		createModuleDescription('vs/platform/files/node/watcher/unix/watcherApp', []),
 		createModuleDescription('vs/platform/files/node/watcher/nsfw/watcherApp', []),
 
-		createModuleDescription('vs/platform/terminal/node/ptyHost/ptyHostMain', []),
+		createModuleDescription('vs/platform/terminal/node/ptyHostMain', []),
 
 		createModuleDescription('vs/workbench/services/extensions/node/extensionHostProcess', []),
 	];
